+++ NEWS.md
@@ -8,7 +8,7 @@
   
   Set `truth_source` to be optional when the user provides `truth_data`. However, it is still needed when `show_caption = TRUE`.
   
-  Remove validation for `model` column in user-proveded `truth_data`.
+  Remove format validation for `model` column in user-provided `truth_data`.
 
 ## covidHubUtils 0.1.1